requirements.txt
langchain_community
langchain_core
langchain_ollama
langgraph
plexapi
